producer/requirements.txt
kafka-python==2.0.2
pandas==2.2.3
requests==2.32.3
python-dateutil==2.9.0.post0
